requirements.txt
fastapi~=0.115.11
uvicorn~=0.34.0
starlette~=0.46.0
python-decouple~=3.6
python-dotenv~=1.0.1
pydantic~=2.10.6
dotenv~=0.9.9
SQLAlchemy~=2.0.38
psycopg2-binary~=2.9.10 # todo : remove package in production mode
asyncpg~=0.30.0
alembic~=1.15.1
PyJWT~=2.10.1
python-multipart~=0.0.7
aiofiles~=24.1.0
PyNaCl~=1.5.0